Add files via upload
The business problem is substantial: out of 7,032 customers, 1,869 churn annually, resulting in a yearly revenue loss of $18,690,000. The goal is to mitigate this loss by building a machine learning (ML) model to predict customer churn and implement strategies to reduce revenue losses by 73%.
The proposed solution involves developing an ML model capable of accurately predicting churn risk. With this model, the company can identify approximately 1,000 customers at high risk of churn. Out of these predictions, 600 are accurate, allowing proactive measures to be taken. By preventing churn in 500 of these cases (83%), the company protects $5,000,000 in revenue.
The overall impact is significant, with a reduction in revenue loss of 73.2%. This underscores the effectiveness of the ML model and the proactive strategies implemented based on its predictions. By leveraging predictive analytics and targeted interventions, the company can significantly improve customer retention and financial performance.